removed extra reducer
