--- src/redux/actions/userActions.js
@@ -1,10 +1,0 @@
-import { REGISTER } from "../constants/users-types";
-
-export function register(payload) {
-    console.log(payload)
-    return { type: REGISTER, payload }
-};
-
-// export const userActions = {
-//     register
-// };
