+++ src/redux/constants/auth-types.js
@@ -1,2 +1,3 @@
 export const LOGIN = "LOGIN";
 export const FOUND_BAD_WORD = "FOUND_BAD_WORD"
+export const REGISTER = "REGISTER";
