--- src/redux/constants/users-types.js
@@ -1,1 +1,0 @@
-export const REGISTER = "REGISTER";
