+++ src/redux/reducers/rootReducer.js
@@ -1,10 +1,8 @@
 import { combineReducers } from 'redux'
 import authReducer from './authReducer'
-import userReducer from './userReducer'
 
 const rootReducer =  combineReducers({
-    authReducer,
-    userReducer
+    authReducer
 })
 
 export default rootReducer;
